Remember Logon, File Writing, and GUI Changes
Program finally saves username and password in a file called "dep.dat"
in the directory of the scraper file. WARNING: The logon information is
stored in a format which could easily be read by anyone so keep it
secret and don't check the box on a public computer. Also adjusted some
GUI spacing and methods.