Optimizing, Reorganizing, and Prettier Output
Fixed imports (again). I had the import for an earlier BeautifulSoup but
that is now fixed for version 4. Along with cleaning up a few lines of
code, I used more BS to make a nicely formatted console output of the
specific grades. The .gitignore was updated to not include login files.
Also, the readme was greatly updated.