Changed cross-platform and Fixed Command Problem
I removed all the differences in running the script on windows compared
to other operating systems. I have not got a chance to test this but I
believe it will work. Also, I fixed some code that I changed yesterday
that caused every button to have the same command as the last button.